store results immediately
computing e.g. 2000 results could be expensive. write results to file as
they complete so they are not lost in the face of errors.

the existing --incremental arg can be used to continue.

duplicates in the result jsonl (e.g. success-after-failure) are OK since
the incremental loader algorithm will take the latest/last result per
hash.